rename activation builtins module
